Add PGD attack

+++ few_shot/core.py
@@ -171,7 +171,7 @@ def prepare_nshot_task(n: int, k: int, q: int) -> Callable:
         TODO: Move to arbitrary device
         """
         x, y = batch
-        x = x.double().cuda()
+        x = x.float().cuda()
         # Create dummy 0-(num_classes - 1) label
         y = create_nshot_task_label(k, q).cuda()
         return x, y
